Automating support for humanoid body parts.
Automating the adding of body parts to humanoid races.
Automating the adding of meats of humanoid races to protein slag recipes.
Automating the adding of similar body parts to RecipeDefs of Surgeries. (As long as the Def for the tail have "tail" anywhere in it's defName it will be added for example)
Deleting of manual patches.